Fixed auth routes for displaying buying option
